make soft deleted users and reset password

+++ app/Http/Controllers/UserController.php
@@ -7,12 +7,17 @@ use Illuminate\Http\Request;
 use Inertia\Inertia;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\UserRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Validation\ValidationException;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(){
-        $users = User::orderBy('id')->paginate(10)->through(function($user){
-            $user->role =  $user->getRoleNames()->first();
+        $users = User::withTrashed()->orderBy('id')->paginate(10)->through(function($user){
+            $user->role = $user->getRoleNames()->first();
+            $user->is_deleted = $user->trashed();
             return $user;
         });
 
@@ -83,11 +88,44 @@ class UserController extends Controller
     }
 
     public function destroy(User $user){
-        $role = $user->getRoleNames()->first();
-        $user->removeRole($role);
+        // $this->isAuthorized($user);
+
+        $this->authorize('delete', $user);
 
         $user->delete();
 
         return redirect()->back()->withSuccess('Η Διαγραφή του χρήστη έγινε με επιτυχία.');
     }
+
+    public function restore_user($id){
+        $user = User::onlyTrashed()->find($id);
+
+        $this->authorize('restore', $user);
+
+        $user->restore();
+
+        return redirect()->back()->withSuccess('Η Επαναφορά του χρήστη έγινε με επιτυχία.');
+    }
+
+    public function final_deleted($id){
+        $user = User::onlyTrashed()->find($id);
+
+        $this->authorize('forceDelete', $user);
+
+        $role = $user->getRoleNames()->first();
+        $user->removeRole($role);
+
+        $user->forceDelete();
+
+        return redirect()->back()->withSuccess('Η οριστική διαγραφή του χρήστη έγινε με επιτυχία.');
+    }
+
+
+    // protected function isAuthorized(User $user){
+    //     if ($user->getRoleNames()->first() == 'Διαχειριστής' && $user->email == 'admin@admin.com') {
+    //         throw ValidationException::withMessages([
+    //             'unauthorizedAction' => __('validation.action'),
+    //         ]);
+    //     }
+    // }
 }
